% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_create_statement_helpers.R
\name{calculateCharSize}
\alias{calculateCharSize}
\title{Get size of character columns}
\usage{
calculateCharSize(col)
}
\arguments{
\item{col}{}
}
\value{
integer size
}
\description{
Get size of character columns
}
