% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_create_statement_helpers.R
\name{colToRedshiftType}
\alias{colToRedshiftType}
\title{Get Redshift Data Type From R Column}
\usage{
colToRedshiftType(col, compression)
}
\arguments{
\item{col}{from a R data frame}

\item{compression}{logical flag}
}
\value{
string with Redshift data type
}
\description{
Get Redshift Data Type From R Column
}
