% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_create_statement_helpers.R
\name{getRedshiftTypesForDataFrame}
\alias{getRedshiftTypesForDataFrame}
\title{Get Redshift Data Types From a Data Frame}
\usage{
getRedshiftTypesForDataFrame(df, compression)
}
\arguments{
\item{df}{The data frame}

\item{compression}{logical flag}
}
\value{
list of data types as characters
}
\description{
Get Redshift Data Types From a Data Frame
}
