% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbcReConnect.R
\name{odbcReConnect}
\alias{odbcReConnect}
\title{If ODBC connection not functioning, re-connect.}
\usage{
odbcReConnect(con, on_table = dbplyr::in_schema("pg_catalog", "pg_type"))
}
\arguments{
\item{con}{An established DBI/odbc connection object}

\item{on_table}{The DB table, possibly in a schema, to use for test.}
}
\value{
The original, or re-estbalished, connection.
}
\description{
Attempts to create a dplyr::tbl() object. If that fails, then odbc::dbConnect() to
re-connect.
}
\details{
Note: \code{odbc::IsValid} does not check the connection is actually still live, which
is why we use \code{dplyr::tbl()} to test.

Will throw error \emph{Could not re-establish connection.}, if new connection is not
\code{identical()} to original.
}
