% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_collapse.R
\name{sql_collapse}
\alias{sql_collapse}
\title{Collapse a Multi-line SQL Query}
\usage{
sql_collapse(query)
}
\arguments{
\item{query}{string The multi-line SQL query with, perhaps, "--" comments.}
}
\description{
Collapses a well formatted SQL query such as you would get with
\code{readLines()} reading a query created with, say, dbVis into
a single character string. Also any "to end of line" comments are
removed before doing the collapse.
}
\examples{
\dontrun{
q <- readLines('./SQL/qVariablesInQueryExample1.sql')
sql_collapse(q)
}
}

